Restructure project README
The general goal here is to improve visibility
of the example, requirements and installation sections
without removing any critical info.

Adds a Links section to the top of the README, for
quick visibility of ReadTheDocs documentation and
the GT gitter channel. Also adds a Gitter badge.

Anything related to developing GPS was moved to a new
Developers section, including installation instructions
specific to developers, to avoid confusion in the
Installation section for regular users.

Moved Uninstallation to the last subheading of the
Installation section.

Moves PYTHONPATH installation notes to developer section.
These notes apply to `pip install -e .` installations,
which are most common for GeoPySpark developers.
So, it makes sense for the note to live in the
developer section.